requirements.txt
flask==3.0.0
gunicorn==21.2.0
python-telegram-bot==20.7
python-dotenv==1.0.0
